python_embed_service/requirements.txt
sentence-transformers
faiss-cpu
fastapi
uvicorn
pydantic
numpy
python-multipart
requests
